+++ tests/e2e_tests/reference_files/bare.txt
@@ -1,0 +1,2 @@
+A bare py-bugger call makes no changes to your project.
+You must be explicit about what kinds of errors you want to induce in the project.
